ref(Diagram) move MyTab to own File
