+++ src/tool/clients/fmmlxdiagrams/fmmlxdiagram/DiagramViewHeadToolBar.java
@@ -1,4 +1,4 @@
-package tool.clients.fmmlxdiagrams.menus;
+package tool.clients.fmmlxdiagrams.fmmlxdiagram;
 
 
 import java.util.EnumSet;
@@ -29,7 +29,6 @@ import tool.clients.fmmlxdiagrams.DiagramDisplayProperty;
 import tool.clients.fmmlxdiagrams.ReturnCall;
 import tool.clients.fmmlxdiagrams.Note;
 import tool.clients.fmmlxdiagrams.dialogs.ShortcutDialog;
-import tool.clients.fmmlxdiagrams.fmmlxdiagram.FmmlxDiagram;
 import tool.clients.fmmlxdiagrams.graphics.wizard.ConcreteSyntaxWizard;
 import tool.communication.java_to_python.MissingPythonRespondException;
 import tool.communication.java_to_python.PythonFunction;
